+++ tests/cli/test_cli_write.py
@@ -130,9 +130,12 @@ DELETE_KEYS_MD = MappingProxyType(
             "identifiers": {
                 "comicvine": {
                     "key": "145269",
-                    "url": "https://comicvine.gamespot.com/captain-science-1/4000-145269/",
                 }
             },
+            "urls": [
+                "https://comicvine.gamespot.com/captain-science-1/4000-145269/",
+                "https://comicvine.gamespot.com/c/4000-145269/",
+            ],
             "imprint": {"name": "CLIImprint"},
             "issue": {
                 "name": "001",