+++ tests/schemas/test_comet.py
@@ -47,9 +47,9 @@ METADATA = MappingProxyType(
             "identifiers": {
                 "comicvine": {
                     "key": "145269",
-                    "url": "https://comicvine.gamespot.com/c/4000-145269/",
                 }
             },
+            "urls": ["https://comicvine.gamespot.com/c/4000-145269/"],
             "issue": {"name": "1", "number": Decimal(1)},
             "language": "en",
             "bookmark": 4,